cargo: fix build

+++ src/uutils/uutils_cargo.rs
@@ -1,4 +1,5 @@
 #![crate_name = "uutils"]
+#![allow(unstable)]
 
 /*
  * This file is part of the uutils coreutils package.
@@ -85,8 +86,8 @@ use std::collections::hash_map::HashMap;
 static NAME: &'static str = "uutils";
 static VERSION: &'static str = "1.0.0";
 
-fn util_map() -> HashMap<&'static str, fn(Vec<String>) -> int> {
-    let mut map : HashMap<&'static str, fn(Vec<String>) -> int> = HashMap::new();
+fn util_map() -> HashMap<&'static str, fn(Vec<String>) -> isize> {
+    let mut map : HashMap<&'static str, fn(Vec<String>) -> isize> = HashMap::new();
 
     macro_rules! add_util(
         ($guard:meta, $util:expr, $crte:ident) => ( 
@@ -176,7 +177,7 @@ fn util_map() -> HashMap<&'static str, fn(Vec<String>) -> int> {
     map
 }
 
-fn usage(cmap: &HashMap<&'static str, fn(Vec<String>) -> int>) {
+fn usage(cmap: &HashMap<&'static str, fn(Vec<String>) -> isize>) {
     println!("{} {}", NAME, VERSION);
     println!("");
     println!("Usage:");
